fix(solver): split empty-key index access helper
