+++ crates/tsz-solver/src/evaluation/evaluate_rules/mod.rs
@@ -17,6 +17,7 @@ pub mod apparent;
 pub mod conditional;
 pub mod index_access;
 mod index_access_callable;
+mod index_access_empty_key;
 mod index_access_keys;
 mod index_access_object_with_index;
 mod index_access_tuple_literal;
